DrawCall structures changed and completed SoftBatch implementation
Changes:
-DrawCall struct generally improved and repurposed to serve as a base for other similar drawcall-like structs
-SoftBatches fully implemented with working shaders and infrastructure in Renderer2D.
--Suported per-instance rotation in radians, position and sprite index

Note: This is likely a problem for any game engines willing to use this, but the positions of the sprites in the soft batches are anchored around the point at which the SoftBatch draw call is made. So if you want to render a bunch of things in a soft batch without extra calculaion, you should just make the DrawCall at the origin (0, 0)

+++ OpenGl-2D-Renderer-Revamped/common/standard_quad.h
@@ -39,11 +39,11 @@ struct StandardQuad {
 
 public:
 
-	unsigned int  g_stdVertexCoordArray[8] = {
+	float  g_stdVertexCoordArray[8] = {
 		0,										0,
-		0,										(unsigned)m_StandardSpritePixelLength,
-		(unsigned)m_StandardSpritePixelLength,	(unsigned)m_StandardSpritePixelLength,
-		(unsigned)m_StandardSpritePixelLength,	0
+		0,									(float)m_StandardSpritePixelLength,
+		(float)m_StandardSpritePixelLength,	(float)m_StandardSpritePixelLength,
+		(float)m_StandardSpritePixelLength,	0
 	};
 
 	float g_stdTexCoordArray[8] = {
